requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
fastapi==0.118.0
h11==0.16.0
idna==3.10
lxml==6.0.1
pydantic==2.11.9
pydantic_core==2.33.2
python-docx==1.2.0
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
